requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
filelock==3.0.12
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
pipenv==2020.11.15
pycparser==2.20
PyJWT==2.1.0
pypi==2.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.4.4
virtualenv-clone==0.5.4
